# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Green Diamond",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "summary": "Mostrar categorias",
                "responses": {
                    "200": {
                        "description": "Mostrar todos las categorias."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/admin/products": {
            "get": {
                "summary": "Mostrar productos",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los productos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}